packages/cli: common yarn install for e2e tests + point to dist
