+++ packages/cli/src/commands/create-plugin/createPlugin.ts
@@ -28,7 +28,7 @@ import {
 } from '../../lib/codeowners';
 import { paths } from '../../lib/paths';
 import { version } from '../../lib/version';
-import { Task, templatingTask } from '../../lib/tasks';
+import { Task, templatingTask, installWithLocalDeps } from '../../lib/tasks';
 const exec = promisify(execCb);
 
 async function checkExists(rootDir: string, id: string) {
@@ -141,7 +141,9 @@ async function cleanUp(tempDir: string) {
 }
 
 async function buildPlugin(pluginFolder: string) {
-  const commands = ['yarn install', 'yarn build'];
+  await installWithLocalDeps(paths.targetRoot);
+
+  const commands = ['yarn tsc', 'yarn build'];
   for (const command of commands) {
     await Task.forItem('executing', command, async () => {
       process.chdir(pluginFolder);
